+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Added] custom field to TracableInfo
 * [Added] fold support
+* [Added] color setting
 
 ## [v0.2.1](https://github.com/dalance/nom-tracable/compare/v0.2.0...v0.2.1) - 2019-07-29